tmrctr: Fix compilation errors with -Werror=conversion
This patch fixes compilation errors that may arise when
-Werror=conversion flag is enabled in compiler flags.

Also updates version and corresponding doxygen tags.


Acked-for-series:  Mubin  Sayyed <mubin.usman.sayyed@xilinx.com>

+++ XilinxProcessorIPLib/drivers/tmrctr/src/xtmrctr_intr.c
@@ -1,5 +1,5 @@
 /******************************************************************************
-* Copyright (C) 2002 - 2020 Xilinx, Inc.  All rights reserved.
+* Copyright (C) 2002 - 2021 Xilinx, Inc.  All rights reserved.
 * SPDX-License-Identifier: MIT
 ******************************************************************************/
 
@@ -7,7 +7,7 @@
 /**
 *
 * @file xtmrctr_intr.c
-* @addtogroup tmrctr_v4_7
+* @addtogroup tmrctr_v4_8
 * @{
 *
 * Contains interrupt-related functions for the XTmrCtr component.
@@ -23,6 +23,8 @@
 *		      definitions.
 * 2.03a rvo  11/30/10 Added check to see if interrupt is enabled before further
 *		      processing for CR 584557.
+* 4.8   dp   02/12/21 Fix compilation errors that arise when -Werror=conversion
+*                     is enabled in compilation flags.
 * </pre>
 *
 ******************************************************************************/
@@ -163,7 +165,7 @@ void XTmrCtr_InterruptHandler(void *InstancePtr)
 						 * interrupt to be acknowledged
 						 */
 						ControlStatusReg &=
-							~XTC_CSR_ENABLE_TMR_MASK;
+							(u32)~XTC_CSR_ENABLE_TMR_MASK;
 
 						XTmrCtr_WriteReg(
 							TmrCtrPtr->BaseAddress,
